update lan 1

+++ src/main/java/com/example/finalproject/domain/Skill.java
@@ -1,7 +1,7 @@
 package com.example.finalproject.domain;
 
 public class Skill {
-    private int id;
+    private Long id;
     private String name;
     private int level;
     private int yearExperience;
@@ -12,7 +12,7 @@ public class Skill {
     public Skill() {
     }
 
-    public Skill(int id, String name, int level, int yearExperience, int monthExperience, String description, Employee employee) {
+    public Skill(Long id, String name, int level, int yearExperience, int monthExperience, String description, Employee employee) {
         this.id = id;
         this.name = name;
         this.level = level;
@@ -22,11 +22,11 @@ public class Skill {
         this.employee = employee;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
